personal website created using Streamlit and Python
The app features various sections, including an introduction, a description of my skills, a portfolio of projects, and a contact form. It uses Lottie animations to enhance user engagement and integrates a form submission service for contacting me
Features:
Introduction: A brief overview of who I am and my professional background.
Skills: An outline of my skills and what I do, with links to relevant content.
Projects: A showcase of my projects, including tutorials and example applications.
Contact Form: A simple form for visitors to get in touch with me, implemented using Form Submit.